refactor: move config to it's own mod
